refactor(random): rename randx package to random
Descriptive name for the crypto-secure entropy helpers; 'random' has no
stdlib collision. Updated consumers: token, password, secret; id test
comment. Error-prefix sentinels now use 'random:'.
